Move init.gradle to resources
